Support child age in search pricing via global age bands
Child age was not priced: occupancy keyed only on (adults, children) count,
so a 6yo and a 15yo cost the same. Fixed with platform-wide age bands so
multi-hotel single-scan search still works.

- New child_age_bands (global): infant 0-1 (excluded, handled by max_infants),
  young 2-6, child 7-12, teen 13-17; 18+ = adult. Hotel/contract child
  policies map into these bands at load time (PHP).
- occupancy_options now carries per-band child counts (band1/2/3_qty) and a
  precomputed global occ_key = adults*1000 + b1*100 + b2*10 + b3.
- search_daily.occupancy_id is that occ_key (widened to INT). Rebuild writes
  o.occ_key. Search resolves child ages -> bands -> occ_key as :occ.
- Verified on the 16.4M-row table: 2 adults + kids 6,8 (occ 2110) vs 15,16
  (occ 2002) return different correct prices via the same index-only
  idx_hotel path.

Docs: new section 3b (child ages), reference seed for child_age_bands.

+++ inventory/database/02_schema_search_cache.sql
@@ -25,10 +25,13 @@ CREATE TABLE search_daily (
   room_id        INT UNSIGNED     NOT NULL,
   rate_plan_id   INT UNSIGNED     NOT NULL,
   board_type_id  TINYINT UNSIGNED NOT NULL,   -- фильтр по типу питания
-  -- ГЛОБАЛЬНЫЙ ключ размещения = adults*100+children (room-independent!).
-  -- В отличие от per-room occupancy_options.id, одинаков во ВСЕХ отелях,
-  -- поэтому :occ фильтрует сразу сотни отелей одним значением.
-  occupancy_id   SMALLINT UNSIGNED NOT NULL,
+  -- ГЛОБАЛЬНАЯ подпись размещения occ_key = adults*1000 + b1*100 + b2*10 + b3,
+  -- где b1..b3 — число детей в каждом ВОЗРАСТНОМ бэнде (child_age_bands).
+  -- Учитывает возраст детей и одинакова во ВСЕХ отелях (бэнды глобальные),
+  -- поэтому :occ фильтрует сотни отелей одним значением, а цена в строке
+  -- уже посчитана под этот возрастной состав. INT — с запасом на расширение
+  -- числа бэндов (при 4+ бэндах кодировка выходит за SMALLINT).
+  occupancy_id   INT UNSIGNED     NOT NULL,
   adults         TINYINT UNSIGNED NOT NULL,   -- денормализовано (для вывода/отладки)
   children       TINYINT UNSIGNED NOT NULL,
   max_infants    TINYINT UNSIGNED NOT NULL DEFAULT 0,  -- вместимость по младенцам (room)
